🤖 Refactor: 이중 모달 구조 개선
-모달 위에 모달이 표시되던 부분을 개선하여, 한 번에 하나의 모달만 화면에 표시되도록 수정하였습니다
